requirements.txt
# requirements.txt

streamlit
pandas
opencv-python-headless
pyspellchecker
pytesseract
Pillow
torch
torchvision
transformers
sentencepiece
matplotlib